+++ oap-server/server-health-checker/src/main/java/org/apache/skywalking/oap/server/health/checker/provider/HealthCheckerProvider.java
@@ -26,7 +26,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.oap.server.core.CoreModule;
 import org.apache.skywalking.oap.server.core.server.HTTPHandlerRegister;
@@ -100,16 +99,22 @@ public class HealthCheckerProvider extends ModuleProvider {
     @Override public void notifyAfterCompleted() throws ServiceNotProvidedException, ModuleStartException {
         ses.scheduleAtFixedRate(() -> {
             StringBuilder unhealthyModules = new StringBuilder();
-            score.set(Stream.ofAll(collector.collect())
-                    .flatMap(metricFamily -> metricFamily.samples)
-                    .filter(sample -> metricsCreator.isHealthCheckerMetrics(sample.name))
-                    .peek(sample -> {
-                        if (sample.value > 0.0) {
-                            unhealthyModules.append(metricsCreator.extractModuleName(sample.name)).append(",");
-                        }
-                    })
-                    .map(sample -> sample.value)
-                    .collect(Collectors.summingDouble(Double::doubleValue)));
+            AtomicDouble unhealthyModule = new AtomicDouble(0);
+            Stream.ofAll(collector.collect())
+                                        .flatMap(metricFamily -> metricFamily.samples)
+                                        .filter(sample -> metricsCreator.isHealthCheckerMetrics(sample.name))
+                                        .forEach(sample -> {
+                                            if (sample.value < 1) {
+                                                unhealthyModules.append(metricsCreator.extractModuleName(sample.name)).append(",");
+                                                unhealthyModule.updateAndGet(v -> v + 1);
+                                            }
+                                        });
+
+            if (unhealthyModule.get() > 0) {
+                score.set(unhealthyModule.get());
+            } else {
+                score.set(0);
+            }
             details.set(unhealthyModules.toString());
             },
             2, config.getCheckIntervalSeconds(), TimeUnit.SECONDS);