+++ docs/en/changes/changes.md
@@ -35,6 +35,12 @@
 * refactor: implement OTEL handler with SPI for extensibility.
 * chore: add `toString` implementation for `StorageID`.
 * chore: add a warning log when connecting to ES takes too long.
+* Fix the query time range in the metadata API.
+* OAP gRPC-Client support `Health Check`.
+* [Break Change] `health_check_xx` metrics make response 1 represents healthy, 0 represents unhealthy.
+* Bump up grpc to 1.70.0.
+* BanyanDB: support new Index rule type `SKIPPING/TREE`, and update the record `log`'s `trace_id` indexType to `SKIPPING`
+* BanyanDB: remove `index-only` from tag setting.
 
 #### UI
 
@@ -52,6 +58,11 @@
 * BanyanDB: Add `Data Lifecycle Stages(Hot/Warm/Cold)` documentation.
 * Add `SWIP-9 Support flink monitoring`.
 * Fix `Metrics Attributes` menu link.
+* Implement the Status API on Settings page.
+* Fix: Add the prefix for http url.
+* Enhance the async-profiling duration options.
+* Enhance the TTL Tab on Setting page.
+* Fix the snapshot charts in alarm page.
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/230?closed=1)